chmod, install: move mode parsing into uucore
