+++ src/uucore/Cargo.toml
@@ -17,13 +17,14 @@ fs = ["libc"]
 utf8 = []
 encoding = ["data-encoding"]
 parse_time = []
+mode = ["libc"]
 utmpx = ["time", "libc"]
 process = ["libc"]
 signals = []
 entries = ["libc"]
 wide = []
 utsname = ["libc"]
-default = ["fs", "libc", "utf8", "utsname", "encoding", "parse_time", "utmpx", "process", "entries", "signals", "wide"]
+default = ["fs", "libc", "utf8", "utsname", "encoding", "parse_time", "mode", "utmpx", "process", "entries", "signals", "wide"]
 
 [lib]
 path = "lib.rs"
