+++ src/uucore/lib.rs
@@ -17,6 +17,8 @@ pub mod utf8;
 pub mod encoding;
 #[cfg(feature = "parse_time")]
 pub mod parse_time;
+#[cfg(feature = "mode")]
+pub mod mode;
 
 #[cfg(all(unix, not(target_os = "fuchsia"), feature = "utmpx"))]
 pub mod utmpx;
